docs: convert VerticalLineChart example to TypeScript (.tsx)
